WIP - major refactoring.
