+++ case_conversion/__init__.py
@@ -1,6 +1,6 @@
-from __future__ import absolute_import
+# from __future__ import absolute_import
 
-from .case_conversion import (
-    camelcase, pascalcase, snakecase, dashcase, kebabcase, spinalcase,
-    constcase, screaming_snakecase, dotcase, separate_words, slashcase,
-    backslashcase)
+# from .case_conversion import (
+#     camelcase, pascalcase, snakecase, dashcase, kebabcase, spinalcase,
+#     constcase, screaming_snakecase, dotcase, separate_words, slashcase,
+#     backslashcase)
